# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank API
  version: 1.0.0
  description: REST API banking service with JWT auth, accounts, cards, transfers, credits and analytics.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /health:
    get:
      tags: [System]
      security: []
      summary: Health check
      responses:
        "200":
          description: Service is healthy
  /register:
    post:
      tags: [Auth]
      security: []
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Email or username already exists
  /login:
    post:
      tags: [Auth]
      security: []
      summary: Login and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT token and user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /me:
    get:
      tags: [Auth]
      summary: Return current authenticated user
      description: Use the token from /login in Swagger Authorize. Paste only the JWT token value.
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Missing or invalid token
  /accounts:
    get:
      tags: [Accounts]
      summary: List current user's accounts
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      tags: [Accounts]
      summary: Create RUB account
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{accountId}/deposit:
    post:
      tags: [Accounts]
      summary: Deposit money to own account
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{accountId}/withdraw:
    post:
      tags: [Accounts]
      summary: Withdraw money from own account
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{accountId}/predict:
    get:
      tags: [Analytics]
      summary: Predict account balance for up to 365 days
      parameters:
        - $ref: "#/components/parameters/AccountId"
        - name: days
          in: query
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 365
      responses:
        "200":
          description: Balance prediction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalancePrediction"
  /cards:
    get:
      tags: [Cards]
      summary: List current user's cards
      responses:
        "200":
          description: Cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      tags: [Cards]
      summary: Issue virtual card for account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardRequest"
      responses:
        "201":
          description: Created card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /transfer:
    post:
      tags: [Transfers]
      summary: Transfer money from own account to another account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "200":
          description: Transfer accepted
  /credits:
    post:
      tags: [Credits]
      summary: Create credit with annuity payment schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditRequest"
      responses:
        "201":
          description: Created credit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credit"
  /credits/{creditId}/schedule:
    get:
      tags: [Credits]
      summary: Get credit payment schedule
      parameters:
        - name: creditId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Payment schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentSchedule"
  /analytics:
    get:
      tags: [Analytics]
      summary: Current month financial analytics
      responses:
        "200":
          description: Monthly analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyAnalytics"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    RegisterRequest:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        username:
          type: string
          example: alice_demo
        password:
          type: string
          format: password
          example: password123
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          format: password
          example: password123
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    CreateAccountRequest:
      type: object
      properties:
        currency:
          type: string
          enum: [RUB]
          default: RUB
    AmountRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
          example: 1000
    TransferRequest:
      type: object
      required: [from_account_id, to_account_id, amount]
      properties:
        from_account_id:
          type: integer
          format: int64
        to_account_id:
          type: integer
          format: int64
        amount:
          type: number
          format: double
          example: 2500
    CreateCardRequest:
      type: object
      required: [account_id]
      properties:
        account_id:
          type: integer
          format: int64
    CreateCreditRequest:
      type: object
      required: [account_id, principal, term_months]
      properties:
        account_id:
          type: integer
          format: int64
        principal:
          type: number
          format: double
          example: 120000
        term_months:
          type: integer
          example: 12
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        username:
          type: string
        created_at:
          type: string
          format: date-time
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        number:
          type: string
        currency:
          type: string
        balance:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        account_id:
          type: integer
          format: int64
        number:
          type: string
          example: 2202********1237
        expiry:
          type: string
          example: 12/29
        last4:
          type: string
    Credit:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        account_id:
          type: integer
          format: int64
        principal:
          type: number
          format: double
        annual_rate:
          type: number
          format: double
        term_months:
          type: integer
        monthly_payment:
          type: number
          format: double
        status:
          type: string
        next_payment_at:
          type: string
          format: date-time
    PaymentSchedule:
      type: object
      properties:
        id:
          type: integer
          format: int64
        credit_id:
          type: integer
          format: int64
        due_date:
          type: string
          format: date-time
        amount:
          type: number
          format: double
        paid:
          type: boolean
        paid_at:
          type: string
          format: date-time
          nullable: true
        penalty_applied:
          type: boolean
    MonthlyAnalytics:
      type: object
      properties:
        income:
          type: number
          format: double
        expense:
          type: number
          format: double
        credit_load:
          type: number
          format: double
        transactions:
          type: integer
          format: int64
    BalancePrediction:
      type: object
      properties:
        account_id:
          type: integer
          format: int64
        days:
          type: integer
        current_balance:
          type: number
          format: double
        planned_payments:
          type: number
          format: double
        predicted_balance:
          type: number
          format: double
    ErrorResponse:
      type: object
      properties:
        error:
          type: string